requirements.txt
fastapi==0.104.1
uvicorn
pydantic==1.10.13
python-jose[cryptography]
python-multipart
python-dotenv
bcrypt
pytest
pytest-asyncio
httpx
redis
firebase-admin
google-auth 